build fix on windows, and fixed a warning.

+++ trunk/psycle-core/src/psycle/core/sequence.cpp
@@ -255,7 +255,7 @@ Sequence::Sequence() {
 	master_pattern_ = new Pattern();
 	// make pattern length endless
 	master_pattern_->timeSignatures().clear();
-	master_pattern_->timeSignatures().push_back(TimeSignature(std::numeric_limits<double>::max()));
+	master_pattern_->timeSignatures().push_back(TimeSignature(std::numeric_limits<float>::max()));
 	master_pattern_->setName("master");
 	master_pattern_->setID(-1);
 	Add(master_pattern_);
